lnm-sdk: refactor `shared` mod, clippy

+++ lnm-sdk/src/shared/models/margin.rs
@@ -218,7 +218,7 @@ impl TryFrom<u64> for Margin {
             return Err(MarginValidationError::TooLow);
         }
 
-        Ok(Self(value as u64))
+        Ok(Self(value))
     }
 }
